tx payment - file reorg
