+++ framework/base/src/types/interaction/mod.rs
@@ -32,9 +32,6 @@ mod tx_env_sc;
 mod tx_from;
 mod tx_gas;
 mod tx_payment;
-mod tx_payment_egld;
-mod tx_payment_egld_value;
-mod tx_payment_normalize;
 mod tx_proxy;
 mod tx_result_handler;
 mod tx_rh_list;
@@ -73,9 +70,6 @@ pub use tx_env_sc::*;
 pub use tx_from::*;
 pub use tx_gas::*;
 pub use tx_payment::*;
-pub use tx_payment_egld::*;
-pub use tx_payment_egld_value::TxEgldValue;
-pub use tx_payment_normalize::TxPaymentNormalize;
 pub use tx_proxy::*;
 pub use tx_result_handler::*;
 pub use tx_rh_list::*;
