+++ framework/base/src/types/interaction/tx_payment.rs
@@ -1,14 +1,55 @@
+mod tx_payment_egld;
+mod tx_payment_egld_value;
+mod tx_payment_multi_esdt;
+mod tx_payment_none;
+mod tx_payment_normalize;
+mod tx_payment_other;
+mod tx_payment_single_esdt;
+
+pub use tx_payment_egld::{Egld, EgldPayment};
+pub use tx_payment_egld_value::TxEgldValue;
+pub use tx_payment_normalize::TxPaymentNormalize;
+
 use crate::{
     api::ManagedTypeApi,
     contract_base::SendRawWrapper,
-    imports::{BigUint, ManagedBuffer},
     types::{
-        EgldOrEsdtTokenPayment, EgldOrMultiEsdtPayment, EsdtTokenPayment, ManagedAddress,
-        MultiEsdtPayment,
+        BigUint, EgldOrEsdtTokenPayment, EgldOrMultiEsdtPayment, EsdtTokenPayment, ManagedAddress,
+        ManagedBuffer, MultiEsdtPayment,
     },
 };
 
-use super::{Egld, FunctionCall, TxEnv};
+use super::{FunctionCall, TxEnv};
+
+/// Describes a payment that is part of a transaction.
+pub trait TxPayment<Env>
+where
+    Env: TxEnv,
+{
+    fn is_no_payment(&self) -> bool;
+
+    fn perform_transfer_execute(
+        self,
+        env: &Env,
+        to: &ManagedAddress<Env::Api>,
+        gas_limit: u64,
+        fc: FunctionCall<Env::Api>,
+    );
+
+    fn into_full_payment_data(self, env: &Env) -> FullPaymentData<Env::Api>;
+}
+
+/// Marks a payment object that only contains EGLD or nothing at all.
+pub trait TxPaymentEgldOnly<Env>: TxPayment<Env>
+where
+    Env: TxEnv,
+{
+    fn with_egld_value<F, R>(&self, f: F) -> R
+    where
+        F: FnOnce(&BigUint<Env::Api>) -> R;
+
+    fn into_egld_payment(self, env: &Env) -> BigUint<Env::Api>;
+}
 
 #[derive(Clone)]
 pub struct AnnotatedEgldPayment<Api>
@@ -46,199 +87,6 @@ where
         Self {
             egld: None,
             multi_esdt: Default::default(),
-        }
-    }
-}
-
-/// Describes a payment that is part of a transaction.
-pub trait TxPayment<Env>
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool;
-
-    fn perform_transfer_execute(
-        self,
-        env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    );
-
-    fn into_full_payment_data(self, env: &Env) -> FullPaymentData<Env::Api>;
-}
-
-impl<Env> TxPayment<Env> for ()
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool {
-        true
-    }
-
-    fn perform_transfer_execute(
-        self,
-        env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    ) {
-        Egld(BigUint::zero()).perform_transfer_execute(env, to, gas_limit, fc);
-    }
-
-    fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
-        FullPaymentData::default()
-    }
-}
-
-impl<Env> TxPayment<Env> for EsdtTokenPayment<Env::Api>
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool {
-        self.amount == 0u32
-    }
-
-    fn perform_transfer_execute(
-        self,
-        env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    ) {
-        MultiEsdtPayment::from_single_item(self).perform_transfer_execute(env, to, gas_limit, fc);
-    }
-
-    fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
-        FullPaymentData {
-            egld: None,
-            multi_esdt: MultiEsdtPayment::from_single_item(self),
-        }
-    }
-}
-
-impl<Env> TxPayment<Env> for &MultiEsdtPayment<Env::Api>
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool {
-        self.is_empty()
-    }
-
-    fn perform_transfer_execute(
-        self,
-        _env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    ) {
-        let _ = SendRawWrapper::<Env::Api>::new().multi_esdt_transfer_execute(
-            to,
-            self,
-            gas_limit,
-            &fc.function_name,
-            &fc.arg_buffer,
-        );
-    }
-
-    fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
-        FullPaymentData {
-            egld: None,
-            multi_esdt: self.clone(),
-        }
-    }
-}
-
-impl<Env> TxPayment<Env> for MultiEsdtPayment<Env::Api>
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool {
-        self.is_empty()
-    }
-
-    fn perform_transfer_execute(
-        self,
-        env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    ) {
-        (&self).perform_transfer_execute(env, to, gas_limit, fc);
-    }
-
-    fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
-        FullPaymentData {
-            egld: None,
-            multi_esdt: self,
-        }
-    }
-}
-
-impl<Env> TxPayment<Env> for EgldOrEsdtTokenPayment<Env::Api>
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool {
-        self.amount == 0u32
-    }
-
-    fn perform_transfer_execute(
-        self,
-        env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    ) {
-        self.map_egld_or_esdt(
-            (to, fc),
-            |(to, fc), amount| Egld(amount).perform_transfer_execute(env, to, gas_limit, fc),
-            |(to, fc), esdt_payment| esdt_payment.perform_transfer_execute(env, to, gas_limit, fc),
-        )
-    }
-
-    fn into_full_payment_data(self, env: &Env) -> FullPaymentData<Env::Api> {
-        self.map_egld_or_esdt(
-            (),
-            |(), amount| TxPayment::<Env>::into_full_payment_data(Egld(amount), env),
-            |(), esdt_payment| TxPayment::<Env>::into_full_payment_data(esdt_payment, env),
-        )
-    }
-}
-
-impl<Env> TxPayment<Env> for EgldOrMultiEsdtPayment<Env::Api>
-where
-    Env: TxEnv,
-{
-    fn is_no_payment(&self) -> bool {
-        self.is_empty()
-    }
-
-    fn perform_transfer_execute(
-        self,
-        env: &Env,
-        to: &ManagedAddress<Env::Api>,
-        gas_limit: u64,
-        fc: FunctionCall<Env::Api>,
-    ) {
-        match self {
-            EgldOrMultiEsdtPayment::Egld(egld_amount) => {
-                Egld(egld_amount).perform_transfer_execute(env, to, gas_limit, fc)
-            },
-            EgldOrMultiEsdtPayment::MultiEsdt(multi_esdt_payment) => {
-                multi_esdt_payment.perform_transfer_execute(env, to, gas_limit, fc)
-            },
-        }
-    }
-
-    fn into_full_payment_data(self, env: &Env) -> FullPaymentData<Env::Api> {
-        match self {
-            EgldOrMultiEsdtPayment::Egld(egld_amount) => {
-                TxPayment::<Env>::into_full_payment_data(Egld(egld_amount), env)
-            },
-            EgldOrMultiEsdtPayment::MultiEsdt(multi_esdt_payment) => {
-                TxPayment::<Env>::into_full_payment_data(multi_esdt_payment, env)
-            },
         }
     }
 }
